first try

+++ src/main/java/com/tongdao/base/commonDao/userInfoDao.java
@@ -1,11 +1,10 @@
 package com.tongdao.base.commonDao;
 
 import com.tongdao.base.model.User.Userinfo;
-import com.tongdao.base.model.product.cementGrindingAid.Cementgrindingaid;
 import com.tongdao.core.Dao.dao.Daoimpl.Daoimpl;
 import org.springframework.stereotype.Repository;
 
 
 @Repository
-public class userInfoDao extends Daoimpl<Userinfo,Long> {
+public class userInfoDao extends Daoimpl<Userinfo,Integer> {
 }
